Split configure and build
CMake step does not build the code anymore. Build is expected to fail,
but CMake is not. So split them, with CMake failing the action if it
won't succeed and ignoring any failure in the first build attempt.

NOTE: CMRX usually has a two-pass build due to the MPU adjustment step
done after the first build.